+++ src/utils/unpack-command.js
@@ -1,5 +1,8 @@
 import fromBase64 from './string/from-base64';
 
-module.exports = function unpackCmd(str) {
-  return JSON.parse(fromBase64(str));
+module.exports = function unpackCmd(str, base64 = true) {
+  if (base64) {
+    return JSON.parse(fromBase64(str));
+  }
+  return JSON.parse(str);
 };